src/requirements.txt
aiohttp
asyncpg
loguru
pydantic
python-dotenv
rich
websockets